Remove logic from model classes and create engine and segments.evaluator modules to hold the business logic
