+++ tests/identities/test_identities_models.py
@@ -1,207 +1,4 @@
-import pytest
-
-from flag_engine.features.constants import STANDARD
-from flag_engine.features.models import FeatureModel, FeatureStateModel
-from flag_engine.identities.models import IdentityModel, TraitModel
-from tests.conftest import segment_condition_property, segment_condition_string_value
-from tests.helpers import get_environment_feature_state_for_feature
-from tests.identities.fixtures import (
-    empty_segment,
-    segment_conditions_and_nested_rules,
-    segment_multiple_conditions_all,
-    segment_multiple_conditions_any,
-    segment_nested_rules,
-    segment_single_condition,
-    trait_key_1,
-    trait_key_2,
-    trait_key_3,
-    trait_value_1,
-    trait_value_2,
-    trait_value_3,
-)
-
-
-def test_identity_get_feature_state_without_any_override(
-    environment, identity, feature_1
-):
-    # When
-    feature_state = identity.get_feature_state(environment, feature_1.name)
-    # Then
-    assert feature_state.feature == feature_1
-
-
-def test_identity_get_all_feature_states_no_segments(
-    feature_1, feature_2, environment, identity
-):
-    # Given
-    overridden_feature = FeatureModel(id=3, name="overridden_feature", type=STANDARD)
-
-    # set the state of the feature to False in the environment configuration
-    environment.feature_states.append(
-        FeatureStateModel(id=3, feature=overridden_feature, enabled=False)
-    )
-
-    # but True for the identity
-    identity.identity_features = [
-        FeatureStateModel(id=4, feature=overridden_feature, enabled=True)
-    ]
-
-    # When
-    all_feature_states = identity.get_all_feature_states(environment=environment)
-
-    # Then
-    assert len(all_feature_states) == 3
-    for feature_state in all_feature_states:
-        environment_feature_state = get_environment_feature_state_for_feature(
-            environment, feature_state.feature
-        )
-
-        expected = (
-            True
-            if feature_state.feature == overridden_feature
-            else environment_feature_state.enabled
-        )
-        assert feature_state.enabled is expected
-
-
-def test_identity_get_all_feature_states_segments_only(
-    feature_1, feature_2, environment, segment, identity_in_segment
-):
-    # Given
-    # a feature which we can override
-    overridden_feature = FeatureModel(id=3, name="overridden_feature", type=STANDARD)
-
-    # which is given a default value of False in the environment configuration
-    environment.feature_states.append(
-        FeatureStateModel(id=3, feature=overridden_feature, enabled=False)
-    )
-
-    # but overridden to True for identities in the segment
-    environment.segment_overrides.append(
-        FeatureStateModel(
-            id=4, feature=overridden_feature, enabled=True, segment_id=segment.id
-        ),
-    )
-
-    # When
-    all_feature_states = identity_in_segment.get_all_feature_states(
-        environment=environment
-    )
-
-    # Then
-    assert len(all_feature_states) == 3
-    for feature_state in all_feature_states:
-        environment_feature_state = get_environment_feature_state_for_feature(
-            environment, feature_state.feature
-        )
-
-        expected = (
-            True
-            if feature_state.feature == overridden_feature
-            else environment_feature_state.enabled
-        )
-        assert feature_state.enabled is expected
-
-
-def test_identity_get_all_feature_states_with_traits(
-    environment_with_segment_override, identity_in_segment, identity
-):
-    # Given
-    trait_models = TraitModel(
-        trait_key=segment_condition_property, trait_value=segment_condition_string_value
-    )
-    # When
-    all_feature_states = identity_in_segment.get_all_feature_states(
-        environment=environment_with_segment_override, traits=[trait_models]
-    )
-    # Then
-    assert all_feature_states[0].get_value() == "segment_override"
-
-
-@pytest.mark.parametrize(
-    "segment, identity_traits, expected_in_segment",
-    (
-        (empty_segment, [], False),
-        (segment_single_condition, [], False),
-        (
-            segment_single_condition,
-            [TraitModel(trait_key=trait_key_1, trait_value=trait_value_1)],
-            True,
-        ),
-        (segment_multiple_conditions_all, [], False),
-        (
-            segment_multiple_conditions_all,
-            [TraitModel(trait_key=trait_key_1, trait_value=trait_value_1)],
-            False,
-        ),
-        (
-            segment_multiple_conditions_all,
-            [
-                TraitModel(trait_key=trait_key_1, trait_value=trait_value_1),
-                TraitModel(trait_key=trait_key_2, trait_value=trait_value_2),
-            ],
-            True,
-        ),
-        (segment_multiple_conditions_any, [], False),
-        (
-            segment_multiple_conditions_any,
-            [TraitModel(trait_key=trait_key_1, trait_value=trait_value_1)],
-            True,
-        ),
-        (
-            segment_multiple_conditions_any,
-            [TraitModel(trait_key=trait_key_2, trait_value=trait_value_2)],
-            True,
-        ),
-        (
-            segment_multiple_conditions_any,
-            [
-                TraitModel(trait_key=trait_key_1, trait_value=trait_value_1),
-                TraitModel(trait_key=trait_key_2, trait_value=trait_value_2),
-            ],
-            True,
-        ),
-        (segment_nested_rules, [], False),
-        (
-            segment_nested_rules,
-            [TraitModel(trait_key=trait_key_1, trait_value=trait_value_1)],
-            False,
-        ),
-        (
-            segment_nested_rules,
-            [
-                TraitModel(trait_key=trait_key_1, trait_value=trait_value_1),
-                TraitModel(trait_key=trait_key_2, trait_value=trait_value_2),
-                TraitModel(trait_key=trait_key_3, trait_value=trait_value_3),
-            ],
-            True,
-        ),
-        (segment_conditions_and_nested_rules, [], False),
-        (
-            segment_conditions_and_nested_rules,
-            [TraitModel(trait_key=trait_key_1, trait_value=trait_value_1)],
-            False,
-        ),
-        (
-            segment_conditions_and_nested_rules,
-            [
-                TraitModel(trait_key=trait_key_1, trait_value=trait_value_1),
-                TraitModel(trait_key=trait_key_2, trait_value=trait_value_2),
-                TraitModel(trait_key=trait_key_3, trait_value=trait_value_3),
-            ],
-            True,
-        ),
-    ),
-)
-def test_identity_in_segment(segment, identity_traits, expected_in_segment):
-    assert (
-        IdentityModel(
-            identifier="identity",
-            environment_api_key="api-key",
-            identity_traits=identity_traits,
-        ).in_segment(segment)
-        == expected_in_segment
-    )
+from flag_engine.identities.models import IdentityModel
 
 
 def test_composite_key():
